Fork to two instruction sets
Inline scripts and external runtime may require divergent implementations (or
hand optimizations for code size). Since we do not need a bundle time fork, I
split this to two different modules.
