+++ packages/react-dom-bindings/src/server/fizz-instruction-set/ReactDOMFizzInlineCompleteBoundary.js
@@ -1,4 +1,4 @@
-import {completeBoundary} from './ReactDOMFizzInstructionSet';
+import {completeBoundary} from './ReactDOMFizzInstructionSetInlineSource';
 
 // This is a string so Closure's advanced compilation mode doesn't mangle it.
 // eslint-disable-next-line dot-notation
