+++ packages/react-dom-bindings/src/server/fizz-instruction-set/ReactDOMFizzInlineCompleteBoundaryWithStyles.js
@@ -1,4 +1,4 @@
-import {completeBoundaryWithStyles} from './ReactDOMFizzInstructionSet';
+import {completeBoundaryWithStyles} from './ReactDOMFizzInstructionSetInlineSource';
 
 // This is a string so Closure's advanced compilation mode doesn't mangle it.
 // eslint-disable-next-line dot-notation
