+++ packages/react-dom-bindings/src/server/fizz-instruction-set/ReactDOMFizzInlineCompleteSegment.js
@@ -1,4 +1,4 @@
-import {completeSegment} from './ReactDOMFizzInstructionSet';
+import {completeSegment} from './ReactDOMFizzInstructionSetInlineSource';
 
 // This is a string so Closure's advanced compilation mode doesn't mangle it.
 // eslint-disable-next-line dot-notation
